+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+* Don't set `cookies.encrypted[:origin]` when passing `?origin` to `new_session_path`
+* `show_signup`, `force_login` and `state` params are now passed through `new_session_path`.
+
 ## 0.6.1
 
 * Fixed token lookup by searching through requested scopes not through granted scopes. You need to run `$ rails zaikio_oauth_client:install:migrations` to apply latest migrations.